config/storage-buckets/requirements.txt
google-cloud-storage
python-dotenv